Modified the synthesis script to avoid keeping additional .v files in the source directory which may confuse qflow when re-running synthesis from the top;  and modified the call to vlogFanout so that it does not overwrite the original input file but makes an additional output file.

+++ scripts/yosys.sh
@@ -171,6 +171,17 @@ if ( ${usescript} == 0 ) then
 
 cd ${sourcedir}
 
+# If there is a file ${modulename}_mapped.v, move it to a temporary
+# place so we can see if yosys generates a new one or not.  Make
+# sure it does *not* have a .v suffix, or else the scripts that
+# search for a file containing the module may pick it up in favor
+# of the actual source file.
+
+rm -f ${modulename}_mapped.v.orig
+if ( -f ${modulename}_mapped.v ) then
+   mv ${modulename}_mapped.v ${modulename}_mapped.v.orig
+endif
+
 # Check for filelists variable overriding the default.  If it is set,
 # use the value as the filename to get all the verilog source files
 # used by the module.  If not, check for a file ${modulename}.fl and
@@ -589,13 +600,6 @@ endif # Generation of yosys script if ${usescript} == 0
 # Yosys synthesis
 #---------------------------------------------------------------------
 
-# If there is a file ${modulename}_mapped.v, move it to a temporary
-# place so we can see if yosys generates a new one or not.
-
-if ( -f ${modulename}_mapped.v ) then
-   mv ${modulename}_mapped.v ${modulename}_mapped_orig.v
-endif
-
 echo "Running yosys for verilog parsing and synthesis" |& tee -a ${synthlog}
 # If provided own script yosys call that, otherwise call yosys with generated script.
 if ( ${usescript} == 1 ) then
@@ -616,14 +620,14 @@ if ( !( -f ${modulename}_mapped.v )) then
    echo "Premature exit." |& tee -a ${synthlog}
    echo "Synthesis flow stopped due to error condition." >> ${synthlog}
    # Replace the old verilog file, if we had moved it
-   if ( -f ${modulename}_mapped_orig.v ) then
-      mv ${modulename}_mapped_orig.v ${modulename}_mapped.v
+   if ( -f ${modulename}_mapped.v.orig ) then
+      mv ${modulename}_mapped.v.orig ${modulename}_mapped.v
    endif
    exit 1
 else
    # Remove the old verilog file, if we had moved it
-   if ( -f ${modulename}_mapped_orig.v ) then
-      rm ${modulename}_mapped_orig.v
+   if ( -f ${modulename}_mapped.v.orig ) then
+      rm ${modulename}_mapped.v.orig
    endif
 endif
 
@@ -678,7 +682,7 @@ endif
 set postproc_options="${postproclefpath} ${postproc_options}"
 
 # Switch to synthdir for processing of the RTL verilog netlist
-cp ${modulename}_mapped.v ${synthdir}/${modulename}.v
+mv ${modulename}_mapped.v ${synthdir}
 cd ${synthdir}
 
 #---------------------------------------------------------------------
@@ -688,13 +692,6 @@ cd ${synthdir}
 if ($?nofanout) then
    set nchanged=0
 else
-
-#---------------------------------------------------------------------
-# Make a copy of the original verilog file, as this will be overwritten
-# by the fanout handling process
-#---------------------------------------------------------------------
-
-   cp ${modulename}.v ${modulename}_bak.v
 
 #---------------------------------------------------------------------
 # Check all gates for fanout load, and adjust gate strengths as
@@ -762,13 +759,12 @@ else
       endif
 
       echo "Running vlogFanout" |& tee -a ${synthlog}
-      echo "vlogFanout ${fanout_options} -I ${modulename}_nofanout ${sepoption} ${libertyoption} ${bufoption} tmp.v ${modulename}_sized.v" |& tee -a ${synthlog}
+      echo "vlogFanout ${fanout_options} -I ${modulename}_nofanout ${libertyoption} ${sepoption} ${bufoption} ${modulename}_mapped.v ${modulename}_sized.v" |& tee -a ${synthlog}
       echo "" >> ${synthlog}
 
-      mv ${modulename}.v tmp.v
       ${bindir}/vlogFanout ${fanout_options} \
 		-I ${modulename}_nofanout ${libertyoption} ${sepoption} \
-		${bufoption} tmp.v ${modulename}_sized.v |& tee -a ${synthlog}
+		${bufoption} ${modulename}_mapped.v ${modulename}_sized.v |& tee -a ${synthlog}
 
       set errcond = ${status}
       if ( ${errcond} != 0 ) then
@@ -783,21 +779,35 @@ else
    endif
 endif
 
+#---------------------------------------------------------------------
+# Spot check:  Did vlogFanout produce an error?
+#---------------------------------------------------------------------
+
+if ( !( -f ${modulename}_sized.v || \
+        ( -M ${modulename}_sized.v < -M ${modulename}_mapped.v ))) then
+   echo "vlogFanout failure.  See file ${synthlog} for error messages." \
+	|& tee -a ${synthlog}
+   echo "Premature exit." |& tee -a ${synthlog}
+   echo "Synthesis flow stopped due to error condition." >> ${synthlog}
+   exit 1
+endif
+
 echo "Running vlog2Verilog for antenna cell mapping." |& tee -a ${synthlog}
 echo "vlog2Verilog -c -p -v ${vddnet} -g ${gndnet} ${postproc_options}" \
 	 |& tee -a ${synthlog}
 echo "   -o ${modulename}.v ${modulename}_sized.v" |& tee -a ${synthlog}
 
+rm -f ${modulename}.v
 ${bindir}/vlog2Verilog -c -p -v ${vddnet} -g ${gndnet} ${postproc_options} \
 	-o ${modulename}.v ${modulename}_sized.v >>& ${synthlog}
 
 #---------------------------------------------------------------------
-# Spot check:  Did vlogFanout produce an error?
+# Spot check:  Did vlog2Verilog produce an error?
 #---------------------------------------------------------------------
 
 if ( !( -f ${modulename}.v || \
-        ( -M ${modulename}.v < -M tmp.v ))) then
-   echo "vlogFanout failure.  See file ${synthlog} for error messages." \
+        ( -M ${modulename}.v < -M ${modulename}_sized.v ))) then
+   echo "vlog2Verilog failure.  See file ${synthlog} for error messages." \
 	|& tee -a ${synthlog}
    echo "Premature exit." |& tee -a ${synthlog}
    echo "Synthesis flow stopped due to error condition." >> ${synthlog}
